# Refactor/tsp speedup