--- a/src/algorithms/munkres.cpp
+++ b/src/algorithms/munkres.cpp
@@ -21,9 +21,8 @@ minimum_weight_perfect_matching(const matrix<T>& m) {
 
   // Trivial initial labeling.
   std::unordered_map<index_t, T> labeling_x;
-  std::unordered_map<index_t, T> labeling_y;
+  std::vector<T> labeling_y(m.size(), 0);
   for (index_t i = 0; i < m.size(); ++i) {
-    labeling_y.emplace(i, 0);
     T min_weight = std::numeric_limits<T>::max();
     for (index_t j = 0; j < m.size(); ++j) {
       if (m[i][j] < min_weight) {
@@ -59,11 +58,11 @@ minimum_weight_perfect_matching(const matrix<T>& m) {
     std::vector<T> slack;
     slack.resize(m.size());
     for (index_t y = 0; y < m.size(); ++y) {
-      if (labeling_x.at(unmatched_x) + labeling_y.at(y) == m[unmatched_x][y]) {
+      if (labeling_x.at(unmatched_x) + labeling_y[y] == m[unmatched_x][y]) {
         alternating_tree.emplace(y, unmatched_x);
       }
       slack[y] = m[unmatched_x][y] - labeling_x.at(unmatched_x) -
-                      labeling_y.at(y);
+                      labeling_y[y];
     }
 
     bool augmented_path = false;
@@ -91,7 +90,7 @@ minimum_weight_perfect_matching(const matrix<T>& m) {
           labeling_x.at(x) = labeling_x.at(x) + alpha;
         }
         for (auto const& y : T_set) {
-          labeling_y.at(y) = labeling_y.at(y) - alpha;
+          labeling_y[y] = labeling_y[y] - alpha;
         }
 
         // Updating relevant neighbors in new equality graph and
@@ -102,7 +101,7 @@ minimum_weight_perfect_matching(const matrix<T>& m) {
 
             if (alternating_tree.find(y) == alternating_tree.end()) {
               for (auto const& x : S) {
-                if (labeling_x.at(x) + labeling_y.at(y) == m[x][y]) {
+                if (labeling_x.at(x) + labeling_y[y] == m[x][y]) {
                     alternating_tree.emplace(y, x);
                 }
               }
@@ -137,7 +136,7 @@ minimum_weight_perfect_matching(const matrix<T>& m) {
         for (index_t y = 0; y < m.size(); ++y) {
           T current_value = slack[y];
           T new_value =
-            m[matched_x][y] - labeling_x.at(matched_x) - labeling_y.at(y);
+            m[matched_x][y] - labeling_x.at(matched_x) - labeling_y[y];
           if (new_value < current_value) {
             slack[y] = new_value;
           }