--- a/src/structures/abstract/edge.cpp
+++ b/src/structures/abstract/edge.cpp
@@ -17,14 +17,6 @@ edge<T>::edge(index_t first_vertex, index_t second_vertex, T weight)
     _weight(weight) {
 }
 
-template <class T> index_t edge<T>::get_first_vertex() const {
-  return _first_vertex;
-}
-
-template <class T> index_t edge<T>::get_second_vertex() const {
-  return _second_vertex;
-}
-
 template <class T> bool edge<T>::operator<(const edge& rhs) const {
   return (this->_first_vertex < rhs._first_vertex) or
          ((this->_first_vertex == rhs._first_vertex) and
@@ -34,10 +26,6 @@ template <class T> bool edge<T>::operator<(const edge& rhs) const {
 template <class T> bool edge<T>::operator==(const edge& rhs) const {
   return (this->_first_vertex == rhs._first_vertex) and
          (this->_second_vertex == rhs._second_vertex);
-}
-
-template <class T> T edge<T>::get_weight() const {
-  return _weight;
 }
 
 template class edge<cost_t>;