--- a/src/structures/abstract/edge.h
+++ b/src/structures/abstract/edge.h
@@ -22,15 +22,21 @@ private:
 public:
   edge(index_t first_vertex, index_t second_vertex, T weight);
 
-  index_t get_first_vertex() const { return _first_vertex;};
+  index_t get_first_vertex() const {
+    return _first_vertex;
+  };
 
-  index_t get_second_vertex() const { return _second_vertex;};
+  index_t get_second_vertex() const {
+    return _second_vertex;
+  };
 
   bool operator<(const edge& rhs) const;
 
   bool operator==(const edge& rhs) const;
 
-  T get_weight() const { return _weight;};
+  T get_weight() const {
+    return _weight;
+  };
 };
 
 #endif