--- a/src/structures/abstract/undirected_graph.cpp
+++ b/src/structures/abstract/undirected_graph.cpp
@@ -16,6 +16,11 @@ template <class T> undirected_graph<T>::undirected_graph() {
 template <class T>
 undirected_graph<T>::undirected_graph(const matrix<T>& m) : _size(m.size()) {
   bool matrix_ok = true;
+  _edges.reserve(_size * _size);
+  _adjacency_list.reserve(_size);
+  for (index_t i = 0; i < _size; ++i) {
+      _adjacency_list[i].reserve(_size);
+  }
   for (index_t i = 0; i < _size; ++i) {
     matrix_ok &= (m[i][i] == INFINITE_COST);
     for (index_t j = i + 1; j < _size; ++j) {
@@ -52,7 +57,11 @@ template <class T> std::vector<edge<T>> undirected_graph<T>::get_edges() const {
 template <class T>
 std::unordered_map<index_t, std::list<index_t>>
 undirected_graph<T>::get_adjacency_list() const {
-  return _adjacency_list;
+  std::unordered_map<index_t, std::list<index_t>> result;
+  for (const auto& pair : _adjacency_list) {
+    std::copy(pair.second.begin(), pair.second.end(), std::back_inserter(result[pair.first]));
+  }
+  return result;
 }
 
 template class undirected_graph<cost_t>;