## Issue

Closes #141 

## Tasks

 - [x] Review algorithm changes
 - [x] Update `CHANGELOG.md` (remove if irrelevant)
 - [x] review

## Description

This PR is a series of patches that improve the performance of the TSP heuristic via some microoptimisations. The patches can be applied and reviewed individually, they do not build upon each other. I think the impact on code quality and maintainability is very small and the performance gains are worth including them.

origin/master:
```
../bin/vroom -i ../bin/d1291.json -t 1  1.93s user 0.12s system 99% cpu 2.051 total
{"cost":53459,"unassigned":0,"computing_times":{"loading":142,"solving":1799}}
```

Improve undirected_graph 003cb5883733ecb51ddcf16a247ea48451c134b5:
```
../bin/vroom -i ../bin/d1291.json -t 1  1.57s user 0.10s system 98% cpu 1.691 total
{"cost":53364,"unassigned":0,"computing_times":{"loading":153,"solving":1430}}
```
The getter copied the `std::list`s anyway, so we avoid their creation cost in the constructor.


Inline edge getters f687ed4:
```
../bin/vroom -i ../bin/d1291.json -t 1  1.53s user 0.10s system 98% cpu 1.658 total
{"cost":53364,"unassigned":0,"computing_times":{"loading":149,"solving":1389}}
```

`get_weight()` showed up in some profiles since it is used to sort the edges by weight for Kruskal's algorithm.

Remove bounds checks in relocate 0ae260b:
```
../bin/vroom -i ../bin/d1291.json -t 1  1.46s user 0.09s system 99% cpu 1.560 total
{"cost":53364,"unassigned":0,"computing_times":{"loading":150,"solving":1329}}
```

Adding shortcut to relocate 696aa4b:
```
../bin/vroom -i ../bin/d1291.json -t 1  1.38s user 0.09s system 99% cpu 1.476 total
{"cost":53364,"unassigned":0,"computing_times":{"loading":139,"solving":1254}}
```

I think this change does not change how the relocate-operator works as long as the triangle inequality holds in the matrix. (However, since I am quite aware that it does not hold on e.g. data that OSRM returns, I have not touched the code below to assume that).

Improve datastructures in munkres c545796 :
```
../bin/vroom -i ../bin/d1291.json -t 1  1.37s user 0.08s system 99% cpu 1.454 total
{"cost":53364,"unassigned":0,"computing_times":{"loading":151,"solving":1223}}
```

The implementation of Munkres' Algorithm does work on all indices of a dense matrix. Thus we don't really need sets and maps and could use vector everywhere.

Replace ordered with unordered set 9154fda :
```
../bin/vroom -i ../bin/d1291.json -t 1  1.33s user 0.08s system 99% cpu 1.412 total
{"cost":53362,"unassigned":0,"computing_times":{"loading":159,"solving":1171}}
```

The tree operations were showing up in profiles.
Note: this commit changes the found solutions since munkres relies on the ordering sometimes

Replace labeling_y with vector b3f25d2: 
```
../bin/vroom -i ../bin/d1291.json -t 1  1.20s user 0.09s system 99% cpu 1.292 total
{"cost":53362,"unassigned":0,"computing_times":{"loading":153,"solving":1056}}
```

Replace labeling_x with vector 0674bde :
```
../bin/vroom -i ../bin/d1291.json -t 1  1.13s user 0.07s system 99% cpu 1.203 total
{"cost":53362,"unassigned":0,"computing_times":{"loading":153,"solving":969}}
```

Improve matrix loading time a bit d98a912:
```
../bin/vroom -i ../bin/d1291.json -t 1  1.05s user 0.09s system 99% cpu 1.143 total
{"cost":53362,"unassigned":0,"computing_times":{"loading":107,"solving":958}}
```
Not really related to the TSP heuristic, but eh well, it showed up in profiles.

## Result

I ran the TSP benchmark script with most of the problems from TSPLIB (I didn't yet run it on the larger ones).

`origin/master`:
```
,Gaps,Computing times
Min,0.09,7
First decile,1.2,11
Lower quartile,1.97,18
Median,2.96,70
Upper quartile,3.56,1419
Ninth decile,4.54,4905
Max,7.56,161364
```

This PR:
```
,Gaps,Computing times
Min,0.0,5
First decile,1.01,13
Lower quartile,1.97,19
Median,2.73,35
Upper quartile,3.59,837
Ninth decile,4.34,3362
Max,6.33,110023
```