@jcoupey 2018-07-24
Wow, thanks for taking the time to profile and work on changes, and also for making it easy to review step by step. Overall all nice improvements, just have a few comments/questions.

1. About 003cb58, I have a feeling that we could stick to using vector instead of list in the return of `get_adjacency_list` to avoid the vector to list copy. This would mean rewriting the work on the eulerian path in `christofides.cpp`. But then maybe replacing the constant-time `erase` used on a list by a vector equivalent would ruin the gain anyway, so it's probably best to leave it that way.

2. Question about commit f687ed4: I don't understand how moving the getters to `edge.h` changes anything from the compiler point of view. Is it related to the template?

3. About 0ae260b: I think this is a left-over from the time where `_edges` used to be an `unordered_map`, I was actually not aware of the overhead of `at` over `operator[]`. Not sure if you have a reason to not apply the same change to every occurrence of `_edges.at`? I think it would be safe to do so.

4. About 9154fda: we can also make the switch to `unordered_set` in `greedy_symmetric_approx_mwpm`, as long as order is consistent in the two `for` loops (don't know if this is guaranteed with `unordered_set`?). Returned matchings might change in case of cost ties, due to lookup re-ordering, but this would not change the overall greedy principle.

Also while we're cleaning up left-overs, we could remove redondant variables in this function:

```patch
--- a/src/algorithms/munkres.cpp
+++ b/src/algorithms/munkres.cpp
@@ -195,8 +195,6 @@ greedy_symmetric_approx_mwpm(const matrix<T>& m) {
 
   while (remaining_indices.size() > 0) {
     T min_weight = std::numeric_limits<T>::max();
-    index_t first_chosen_index;
-    index_t second_chosen_index;
     std::set<index_t>::iterator chosen_i;
     std::set<index_t>::iterator chosen_j;
     for (auto i = remaining_indices.begin(); i != remaining_indices.end();
@@ -207,14 +205,12 @@ greedy_symmetric_approx_mwpm(const matrix<T>& m) {
         T current_weight = m[*i][*j];
         if (current_weight < min_weight) {
           min_weight = current_weight;
-          first_chosen_index = *i;
-          second_chosen_index = *j;
           chosen_i = i;
           chosen_j = j;
         }
       }
     }
-    matching.emplace(first_chosen_index, second_chosen_index);
+    matching.emplace(*chosen_i, *chosen_j);
     remaining_indices.erase(chosen_j);
     remaining_indices.erase(chosen_i);
   }
```