@jcoupey 2018-07-24
Results on all 78 TSPLIB instances on the machine I use for benchmarking:

`origin/master`

```
,Gaps,Computing times
Min,0.33,4
First decile,1.37,6
Lower quartile,2.03,8
Median,2.96,66
Upper quartile,3.72,1137
Ninth decile,4.41,19203
Max,7.56,3458318
```

This PR:

```
,Gaps,Computing times
Min,0.0,3
First decile,0.9,5
Lower quartile,1.78,6
Median,2.93,28
Upper quartile,3.54,466
Ninth decile,5.15,6785
Max,7.56,2245798
```

Total computing time is reduced by ~44%! :tada: