@krypt-n 2018-07-24
> Question about commit f687ed4: I don't understand how moving the getters to edge.h changes anything from the compiler point of view. Is it related to the template?

It is not related to the template. On origin/master, when compiling the `std::sort` in kruskal.cpp, the compiler knows that there is a `edge::get_weight()`-method (since it is declared in the included edge.h header), but it does not know how it is implemented (since edge.cpp is compiled separately). Thus it has to use a function call to get the weight member of an edge class. With f687ed4 the compiler can see the actual implementation of the getter in the header and decide that it is small enough to be inlined into the comparison lambda (which can in turn be inlined into the instantiation of the std::sort template). This way the function call in a hot part of the code turns into a simple pointer dereference.

> Not sure if you have a reason to not apply the same change to every occurrence of _edges.at

No specific reason, just wanted to keep the changes small. Done in 90fcb06

> About 003cb58, I have a feeling that we could stick to using vector instead of list in the return of get_adjacency_list to avoid the vector to list copy

This is probably not a big performance issue right now, since the `eulerian_graph` is not a very dense graph.

> Also while we're cleaning up left-overs, we could remove redundant variables in this function:

Done in 2682f58

> don't know if this is guaranteed with unordered_set

It's not stated explicitly anywhere, but I think the documentation of `rehash` implies that the ordering only changes on `rehash`.

> we can also make the switch to unordered_set in greedy_symmetric_approx_mwpm

The matrix passed to `greedy_symmetric_approx_mwpm` does not seem to be very big in the christofides implementation. I can take a look at that if it starts to appear  in profiles for larger tsp instances.