@jcoupey 2018-07-24
Another thing caught my eye: you report a worst gap of +6.33% with this PR, but I'm still getting the +7.56% (on instance `pr76`) on the machine I used to run the benchmarks. Investigating this a bit, I found out that this worst case is also reduced on my laptop. In fact most results actually differ between the other machine and my laptop (both running at d98a912). I checked the behaviour on `master` and some differences also appear, but to a much lesser extent.

My guess is that this is due to differences in the inner ordering behaviour from different standard library versions (the two machines run Ubuntu 16.04 and 18.04 respectively). And this has been all the more stressed in this PR by additional use of `unordered_set` over `set`. What do you think?