@krypt-n 2018-07-24
I pushed two more commits to this branch:
067ac99 adds a `std::vector` for iteration over the elements of `S`. This means that the iteration order should no longer depend on the implementation of the hashmap but on the insertion order in the munkres algorithm.
8cb214f adds a break a inner loop since we only need to emplace something once. (This is closer to the inner loop on origin/master).

> What do you think?

I think we should try to produce consistent results across different machines and standard libraries. This allows reproducible benchmark results and the scenario where a job is assigned to a vehicle on one computer but unassigned on another computer should probably be avoided as soon as people start using vroom to plan stuff that involves real jobs. 

To continue this line of thinking, the result should probably be independent of the ordering of the input and maybe even independent of the ids.