@jcoupey 2018-07-25
On the matter of reproducible results across machines, this PR did not introduce the non-deterministic behaviours, only maybe made them a bit more noticeable. Handling this here seems out of scope so I'd be inclined to merge as is and ticket this question separately.

I this sounds good to you, could you please update the changelog by adding an item in an `[Unreleased]` entry? Don't worry about potential conflicts with current `master`, they'll be easy to fix upon merging.