@jcoupey 2018-07-27
Shall I let you open a new ticket about reproducible solving based on your previous comment, or do you want me to open it?